+++ core/src/algorithms/fast_decompose.rs
@@ -6,7 +6,7 @@ use std::collections::BTreeMap;
 
 use crate::{
     number::{Dimension, Quantity},
-    numeric::Numeric,
+    types::Numeric,
     Number,
 };
