+++ core/src/ast/mod.rs
@@ -2,7 +2,7 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
-use crate::numeric::{Digits, Numeric};
+use crate::types::{Digits, Numeric};
 use chrono_tz::Tz;
 use serde_derive::Serialize;
 use std::fmt;
