+++ core/src/loader/load.rs
@@ -5,9 +5,9 @@
 use super::Context;
 use crate::ast::{BinOpExpr, Def, DefEntry, Defs, Expr};
 use crate::number::{Dimension, Number};
-use crate::numeric::Numeric;
 use crate::runtime::Value;
 use crate::substance::{Properties, Property, Substance};
+use crate::types::Numeric;
 use std::collections::{BTreeMap, BTreeSet};
 use std::rc::Rc;
 use std::sync::Arc;
