+++ core/src/parsing/formula.rs
@@ -3,8 +3,8 @@
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
 use crate::number::{Dimension, Number};
-use crate::numeric::Numeric;
 use crate::substance::{Properties, Property, Substance};
+use crate::types::Numeric;
 use std::collections::BTreeMap;
 use std::iter::Peekable;
 use std::str::{Chars, FromStr};
