+++ core/src/parsing/text_query.rs
@@ -3,8 +3,7 @@
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
 use crate::ast::*;
-use crate::numeric::{Digits, Numeric};
-use crate::types::{BigInt, BigRat};
+use crate::types::{BigInt, BigRat, Digits, Numeric};
 use chrono_tz::Tz;
 use std::iter::Peekable;
 use std::str::Chars;
